restructure project organization and refactor code
