+++ fin_maestro_kin/main.py
@@ -2,9 +2,8 @@ from fastapi import FastAPI, Request
 from starlette.responses import RedirectResponse
 from modules.sentiment_analysis.sentiment_analysis import router as sentiment_analysis_router
 from modules.trend_detector.trend_detector import router as trend_detector_router
-from modules.data_toolkit.nse.nse_indices import router as nse_indices_router
-from modules.data_toolkit.nse.nse_equities import router as nse_eq_router
-from modules.data_toolkit.screener.screener_equities import router as screener_eq_router
+from modules.data_toolkit.nse.nse_operations import  NSEIndices, NSEEquities
+from modules.data_toolkit.screener.screener_equities import ScreenerEquities
 
 app = FastAPI()
 
@@ -18,6 +17,13 @@ async def apidog_docs_redirect(request: Request, call_next):
 
 app.include_router(sentiment_analysis_router)
 app.include_router(trend_detector_router)
-app.include_router(nse_indices_router)
-app.include_router(nse_eq_router)
-app.include_router(screener_eq_router)
+
+screener_eq = ScreenerEquities()
+screener_eq.register_routes(app)
+
+nse_eq = NSEEquities()
+nse_eq.register_routes(app)
+
+nse_indices = NSEIndices()
+nse_indices.register_routes(app)
+
