+++ pyproject.toml
@@ -1,7 +1,7 @@
 [tool.poetry]
 name = "fin-maestro-kin"
 version = "0.2.5"
-description = "Where Data Meets FastAPI Brilliance with Fin-Maestro-Kin – Your All-in-One Finance API."
+description = "Seamless Finance: Docker Deployed APIs for Smart Investments."
 authors = ["DEV_FINWIZ <devjuneja43@gmail.com>"]
 license = "MIT"
 readme = "README.md"
